D8AGE-379: Move token to the proper namespace and fix test coverage.
