+++ tests/src/Endpoint/StatusEndpointTest.php
@@ -7,8 +7,8 @@ namespace OpenEuropa\Tests\CdtClient\Endpoint;
 use GuzzleHttp\Psr7\Response;
 use OpenEuropa\CdtClient\Endpoint\StatusEndpoint;
 use OpenEuropa\CdtClient\Exception\ValidationErrorsException;
+use OpenEuropa\CdtClient\Model\Response\Token;
 use OpenEuropa\CdtClient\Model\Response\ValidationErrors;
-use OpenEuropa\CdtClient\Model\Token;
 use OpenEuropa\Tests\CdtClient\Traits\AssertTestRequestTrait;
 use OpenEuropa\Tests\CdtClient\Traits\ClientTestTrait;
 use OpenEuropa\Tests\CdtClient\Traits\ResponseModelTestTrait;
