+++ tests/src/Endpoint/ValidateEndpointTest.php
@@ -7,10 +7,10 @@ namespace OpenEuropa\Tests\CdtClient\Endpoint;
 use GuzzleHttp\Psr7\Response;
 use OpenEuropa\CdtClient\Endpoint\ValidateEndpoint;
 use OpenEuropa\CdtClient\Exception\ValidationErrorsException;
+use OpenEuropa\CdtClient\Model\Response\Token;
 use OpenEuropa\CdtClient\Model\Response\ValidationErrors;
-use OpenEuropa\CdtClient\Model\Token;
-use OpenEuropa\Tests\CdtClient\Traits\ClientTestTrait;
 use OpenEuropa\Tests\CdtClient\Traits\AssertTestRequestTrait;
+use OpenEuropa\Tests\CdtClient\Traits\ClientTestTrait;
 use OpenEuropa\Tests\CdtClient\Traits\RequestModelTestTrait;
 use PHPUnit\Framework\TestCase;
 use Psr\Http\Message\RequestInterface;
