+++ tests/src/Traits/ResponseModelTestTrait.php
@@ -14,6 +14,7 @@ use OpenEuropa\CdtClient\Model\Response\ReferenceData;
 use OpenEuropa\CdtClient\Model\Response\ReferenceFile;
 use OpenEuropa\CdtClient\Model\Response\ReferenceItem;
 use OpenEuropa\CdtClient\Model\Response\SourceDocument;
+use OpenEuropa\CdtClient\Model\Response\Token;
 use OpenEuropa\CdtClient\Model\Response\Translation;
 
 /**
@@ -23,6 +24,18 @@ use OpenEuropa\CdtClient\Model\Response\Translation;
  */
 trait ResponseModelTestTrait
 {
+    /**
+     * @param array<string, mixed> $data
+     */
+    public function createResponseToken(array $data = []): Token
+    {
+        return (new Token())
+            ->setAccessToken($data['accessToken'] ?? 'TOKEN')
+            ->setExpiresIn($data['expiresIn'] ?? 28799)
+            ->setTokenType($data['tokenType'] ?? 'bearer')
+            ->setRefreshToken($data['refreshToken'] ?? '{\"TokenId\":\"1234567890abcdef1234567890abcdef\",\"Issued\":\"2024-02-21T07:28:37.7644661Z\",\"Expires\":\"2024-02-22T07:28:37.7644661Z\"}');
+    }
+
     /**
      * @param array<string, mixed> $data
      */
